+++ iceoryx_hoofs/test/moduletests/test_cxx_deadline_timer.cpp
@@ -133,8 +133,8 @@ TIMING_TEST_F(DeadlineTimer_test, RemainingTimeCheckIfExpired, Repeat(5), [&] {
 
     TIMING_TEST_ASSERT_TRUE(sut.hasExpired());
 
-    int remainingTime = sut.remainingTime().toMilliseconds();
-    const int EXPECTED_REMAINING_TIME = 0; // the timer is expired the remaining wait time is Zero
+    uint64_t remainingTime = sut.remainingTime().toMilliseconds();
+    const uint64_t EXPECTED_REMAINING_TIME = 0; // the timer is expired the remaining wait time is Zero
     TIMING_TEST_EXPECT_TRUE(remainingTime == EXPECTED_REMAINING_TIME);
 });
 
@@ -145,7 +145,7 @@ TIMING_TEST_F(DeadlineTimer_test, RemainingTimeCheckIfNotExpired, Repeat(5), [&]
 
     TIMING_TEST_ASSERT_FALSE(sut.hasExpired());
 
-    int remainingTime = sut.remainingTime().toMilliseconds();
+    uint64_t remainingTime = sut.remainingTime().toMilliseconds();
     const int PASSED_TIMER_TIME = SLEEPTIME; // Already 10ms passed in sleeping out of 20ms
     const int RANGE_APPROX = 2;              // 2ms arppoximation. This may be lost after arming the timer in execution.
     const int EXPECTED_REMAINING_TIME = PASSED_TIMER_TIME - RANGE_APPROX;
